+++ tests/Framework/Bootloader/Queue/QueueBootloaderTest.php
@@ -83,6 +83,7 @@ final class QueueBootloaderTest extends BaseTestCase
             'interceptors' => [
                 'consume' => [
                     \Spiral\Queue\Interceptor\Consume\ErrorHandlerInterceptor::class,
+                    \Spiral\Queue\Interceptor\Consume\RetryPolicyInterceptor::class,
                 ],
                 'push' => []
             ],